proati-reservas/backend/requirements.txt
flask
mysql-connector-python
gspread
flask-cors
